refactor: 피드백 반영, 오류 수정 * return type (int -> double)

+++ src/main/java/lotto/domain/Rank.java
@@ -24,7 +24,7 @@ public enum Rank {
                 .orElse(MISS);
     }
 
-    public int getWinningMoney() {
+    public double getWinningMoney() {
         return winningMoney;
     }
 }
